+++ HyPER/utils/__init__.py
@@ -1,7 +1,5 @@
-from .settings import Settings
 from .connectivity import getUndirectedEdges
 
 __all__ = [
-    'Settings',
     'getUndirectedEdges'
 ]